atualização css

+++ gastos/src/app/component/mes/month.component.ts
@@ -4,6 +4,7 @@ import {
   OnInit,
   OnDestroy,
   ChangeDetectorRef,
+  ViewEncapsulation,
 } from '@angular/core';
 import { DateService } from '../../service/date.service';
 import { GastoService } from '../../service/gasto.service';
@@ -11,13 +12,17 @@ import { Gasto } from '../../models/gasto.model';
 import { DatePipe } from '@angular/common';
 import { Subscription } from 'rxjs';
 import { take } from 'rxjs/operators';
-import { DespesaService } from '../../service/despesas.service';
+import { LocalService } from '../../service/local.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material/dialog';
+import { MonthService } from '../../service/month.service';
 
 @Component({
   selector: 'app-month',
   templateUrl: '../mes/month.component.html',
   styleUrls: ['../mes/month.component.css'],
   providers: [DatePipe],
+  encapsulation: ViewEncapsulation.None,
 })
 export class MonthComponent implements OnInit, OnDestroy {
   hoje: Date = new Date();
@@ -28,18 +33,18 @@ export class MonthComponent implements OnInit, OnDestroy {
 
   // Map de meses com os nomes e números
   monthNames: { [key: string]: [string, number] } = {
-    jan: ['January', 1],
-    fev: ['February', 2],
-    mar: ['March', 3],
-    abr: ['April', 4],
-    mai: ['May', 5],
-    jun: ['June', 6],
-    jul: ['July', 7],
-    ago: ['August', 8],
-    set: ['September', 9],
-    out: ['October', 10],
-    nov: ['November', 11],
-    dez: ['December', 12],
+    jan: ['Janeiro', 1],
+    fev: ['Fevereiro', 2],
+    mar: ['Masço', 3],
+    abr: ['Abril', 4],
+    mai: ['Maio', 5],
+    jun: ['Junho', 6],
+    jul: ['Julho', 7],
+    ago: ['Agosto', 8],
+    set: ['Setembro', 9],
+    out: ['Outubro', 10],
+    nov: ['Novembro', 11],
+    dez: ['Dezembro', 12],
   };
 
   private subscription = new Subscription();
@@ -48,13 +53,27 @@ export class MonthComponent implements OnInit, OnDestroy {
     private dateService: DateService,
     private gastoService: GastoService,
     private datePipe: DatePipe,
-    private despesaService: DespesaService,
-    private cdr: ChangeDetectorRef
+    private localService: LocalService,
+    private cdr: ChangeDetectorRef,
+    private dialog: MatDialog,
+    private monthService: MonthService
   ) {}
 
   ngOnInit(): void {
     this.subscribeToDateChanges();
     this.subscribeToDespesaUpdates();
+    this.totalDespesas = this.calculateTotalDespesas();
+
+    // Enviar para o HeaderComponent após o valor ser calculado
+    this.sendDespesasToHeader();
+  }
+
+  // Método auxiliar para formatar o valor como moeda (BRL)
+  private formatCurrency(value: number): string {
+    return value.toLocaleString('pt-BR', {
+      style: 'currency',
+      currency: 'BRL',
+    });
   }
 
   ngOnDestroy(): void {
@@ -77,12 +96,25 @@ export class MonthComponent implements OnInit, OnDestroy {
   // Inscreve-se para atualizações de despesas
   private subscribeToDespesaUpdates(): void {
     this.subscription.add(
-      this.despesaService.despesas$.subscribe((despesas) => {
+      this.localService.despesas$.subscribe((despesas) => {
         this.despesas = despesas;
         this.calcularTotalDespesas();
         this.cdr.detectChanges(); // Detectar mudanças após a atualização das despesas
       })
     );
+  }
+
+  // Método para calcular ou obter o valor total de despesas
+  calculateTotalDespesas(): number {
+    // Aqui você pode incluir a lógica real para calcular o total de despesas
+    // Exemplo: recuperar o valor de uma API ou calcular com base em outras variáveis
+    return 1180.59; // Valor de exemplo, pode ser dinâmico
+  }
+
+  // Método para enviar o valor de despesas para o serviço
+  sendDespesasToHeader(): void {
+    const formattedTotal = this.formatCurrency(this.totalDespesas);
+    this.monthService.setDespesasTotal(formattedTotal, this.selectedMonth);
   }
 
   // Obtém o número do mês baseado no nome
@@ -102,6 +134,13 @@ export class MonthComponent implements OnInit, OnDestroy {
     this.gastoService.getDespesas(month).subscribe((despesas: Gasto[]) => {
       this.despesas = despesas; // Não recalcula o status no frontend
       this.calcularTotalDespesas();
+
+      // Desabilitar botões para despesas pagas
+      this.despesas.forEach((despesa) => {
+        if (despesa.status === 'Pago') {
+          despesa.disableButtons = true;
+        }
+      });
     });
   }
 
@@ -111,6 +150,8 @@ export class MonthComponent implements OnInit, OnDestroy {
       (acc, despesa) => acc + despesa.valor,
       0
     );
+    // Enviar o valor total calculado para o HeaderComponent
+    this.sendDespesasToHeader();
   }
 
   // Marca uma despesa como paga
@@ -122,16 +163,20 @@ export class MonthComponent implements OnInit, OnDestroy {
     despesa.tipo = 1; // Tipo 1 indica que a despesa está paga
     despesa.status = 'Pago';
     despesa.cssClass = 'status-pago'; // Atualiza a classe para "status-pago"
+    despesa.disableButtons = true;
+    console.log(`Despesa ${id} paga e botões desabilitados`);
 
     this.gastoService.pagarDespesa(id).subscribe({
       next: () => {
         this.calcularTotalDespesas(); // Atualiza o total de despesas
+        this.cdr.detectChanges(); // Força a detecção de mudanças
       },
       error: (error) => {
         // Se falhar, reverte para o estado anterior
         despesa.tipo = 0; // Tipo 0 para "Pendente"
         despesa.status = 'Pendente';
         despesa.cssClass = 'status-pendente'; // Retorna para a classe "status-pendente"
+        despesa.disableButtons = false; // Reabilita os botões
         console.error('Erro ao pagar despesa:', error);
       },
     });
@@ -176,6 +221,7 @@ export class MonthComponent implements OnInit, OnDestroy {
           // Atribui o status e a classe CSS ao objeto da despesa
           updatedDespesa.status = infoDespesa.status;
           updatedDespesa.cssClass = infoDespesa.cssClass;
+          updatedDespesa.disableButtons = infoDespesa.disableButtons;
 
           const index = this.despesas.findIndex(
             (d) => d.id === updatedDespesa.id
@@ -200,20 +246,49 @@ export class MonthComponent implements OnInit, OnDestroy {
 
   // Remove uma despesa
   removerDespesa(despesa: Gasto): void {
-    this.gastoService
-      .deleteDespesa(despesa.id!)
-      .pipe(take(1))
-      .subscribe({
-        next: () => {
-          this.despesas = this.despesas.filter((d) => d.id !== despesa.id);
-          this.calcularTotalDespesas();
-        },
-        error: (error) =>
-          console.error('Erro ao remover despesa:', error.message || error),
-      });
+    // Abre o diálogo de confirmação
+
+    const vencimentoFormatado = new Date(despesa.vencimento).toLocaleDateString(
+      'pt-BR',
+      {
+        day: '2-digit',
+        month: '2-digit',
+        year: 'numeric',
+      }
+    );
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '390px',
+      data: {
+        message: `despesa: <strong>${despesa.nome}</strong><br>
+        valor: <strong>R$ ${despesa.valor},00</strong><br>
+        vencimento: <strong>${despesa.vencimento}</strong>`,
+      },
+    });
+
+    // Quando o usuário clicar em confirmar, executa a exclusão
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result === 'confirm') {
+        this.gastoService
+          .deleteDespesa(despesa.id!)
+          .pipe(take(1))
+          .subscribe({
+            next: () => {
+              this.despesas = this.despesas.filter((d) => d.id !== despesa.id);
+              this.calcularTotalDespesas();
+            },
+            error: (error) =>
+              console.error('Erro ao remover despesa:', error.message || error),
+          });
+      }
+    });
   }
 
-  getStatus(despesa: Gasto): { status: string; cssClass: string } {
+  getStatus(despesa: Gasto): {
+    status: string;
+    cssClass: string;
+    disableButtons: boolean;
+  } {
     const hoje = new Date();
     let vencimento: Date;
 
@@ -238,20 +313,32 @@ export class MonthComponent implements OnInit, OnDestroy {
 
     // Se o tipo for 1 (Pago)
     if (despesa.tipo === 1) {
-      return { status: 'Pago', cssClass: 'status-pago' };
+      return { status: 'Pago', cssClass: 'status-pago', disableButtons: true };
     }
 
     // Se o vencimento for hoje
     if (vencimentoSemHora.getTime() === hojeSemHora.getTime()) {
-      return { status: 'Vencendo', cssClass: 'status-hoje' };
+      return {
+        status: 'Vencendo',
+        cssClass: 'status-hoje',
+        disableButtons: true,
+      };
     }
 
     // Se o vencimento passou
     if (vencimentoSemHora < hojeSemHora) {
-      return { status: 'Vencido', cssClass: 'status-vencido' };
+      return {
+        status: 'Vencido',
+        cssClass: 'status-vencido',
+        disableButtons: true,
+      };
     }
 
     // Caso contrário, está pendente (vencimento futuro)
-    return { status: 'Pendente', cssClass: 'status-pendente' };
+    return {
+      status: 'Pendente',
+      cssClass: 'status-pendente',
+      disableButtons: true,
+    };
   }
 }
